Extract diplomacy planner module to shrink domain_planners and preserve helper APIs.
This keeps existing diplomacy scoring/selection behavior intact while moving code into a focused module, bringing domain_planners below the issue line-count target for the ongoing refactor slices.
